split crate into modules
