+++ README.md
@@ -37,21 +37,18 @@ use elastic::prelude::*;
 let client = Client::new(RequestParams::default()).unwrap();
 
 // A search request with a freeform body.
-let req = {
-    let query = json!({
-        "query": {
-            "query_string": {
-                "query": "*"
-            }
-        }
-    });
-
-    SearchRequest::for_index("_all", query.to_string())
-};
-
-// Send the request and process the response.
 let res = client.search::<Value>()
-                .request(req)
+                .index("_all")
+                .body({
+                    json!({
+                        "query": {
+                            "query_string": {
+                                "query": "*"
+                            }
+                        }
+                    })
+                    .to_string()
+                })
                 .send()
                 .unwrap();
 
@@ -59,8 +56,6 @@ let res = client.search::<Value>()
 for hit in res.hits() {
     println!("{:?}", hit);
 }
-
-println!("{:?}", res);
 ```
 
 See the [examples](https://github.com/elastic-rs/elastic/tree/master/elastic/examples) folder for complete samples.
